+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Add `timeout` and `timeoutValue` options to
   `SyncMessagePort.receiveMessage()`.
 
+* Add a `closedValue` option to `SyncMessagePort.receiveMessage()`.
+
 ## 1.0.0
 
 * Initial release